refactor: update player version to 5.1.0
