+++ BasicDownload/src/main/java/com/blendvision/player/basic/download/sample/MainViewModel.kt
@@ -1,8 +1,8 @@
 package com.blendvision.player.basic.download.sample
 
 import androidx.lifecycle.ViewModel
-import com.blendvision.player.download.Downloader
-import com.blendvision.player.download.ui.entity.DownloadConfig
+import com.blendvision.player.download.presentation.Downloader
+import com.blendvision.player.download.presentation.entity.DownloadConfig
 import com.blendvision.player.playback.presentation.UniPlayer
 import com.blendvision.player.playback.presentation.entity.OfflineMediaConfig
