+++ BasicDownload/src/main/java/com/blendvision/player/basic/download/sample/factory/MainViewModelFactory.kt
@@ -5,9 +5,11 @@ import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.blendvision.player.basic.download.sample.MainViewModel
-import com.blendvision.player.download.Downloader
-import com.blendvision.player.download.ui.entity.DownloadTrackSelection
-import com.blendvision.player.download.ui.entity.DownloadableTracks
+import com.blendvision.player.common.presentation.entity.log.LogEvent
+import com.blendvision.player.common.presentation.entity.log.LogProperty
+import com.blendvision.player.download.presentation.Downloader
+import com.blendvision.player.download.presentation.entity.DownloadTrackSelection
+import com.blendvision.player.download.presentation.entity.DownloadableTracks
 import com.blendvision.player.playback.presentation.UniPlayer
 import com.blendvision.player.playback.presentation.entity.PlayerConfig
 import com.blendvision.player.playback.presentation.logger.PlayLogger
@@ -36,8 +38,8 @@ class MainViewModelFactory(
             PlayerConfig(
                 license = PLAYER_LICENSE,
                 playLogger = object : PlayLogger {
-                    override fun logEvent(eventName: String, properties: Map<String, Any>) {
-                        Log.d(TAG, "$eventName, $properties")
+                    override fun onLogEvent(logEvent: LogEvent, properties: Map<LogProperty, Any>) {
+                        Log.d(TAG, "${logEvent.name}, ${properties.mapKeys { it.key.name }}")
                     }
                 }
             )
