+++ BasicPlayback/src/main/java/com/blendvision/player/basic/playback/sample/MainViewModel.kt
@@ -2,6 +2,8 @@ package com.blendvision.player.basic.playback.sample
 
 import android.content.Context
 import androidx.lifecycle.ViewModel
+import com.blendvision.player.common.presentation.entity.log.LogEvent
+import com.blendvision.player.common.presentation.entity.log.LogProperty
 import com.blendvision.player.playback.presentation.UniPlayer
 import com.blendvision.player.playback.presentation.entity.MediaConfig
 import com.blendvision.player.playback.presentation.entity.PlayerConfig
@@ -30,7 +32,7 @@ class MainViewModel : ViewModel() {
             playerConfig = PlayerConfig(
                 license = PLAYER_LICENSE,
                 playLogger = object : PlayLogger {
-                    override fun logEvent(eventName: String, properties: Map<String, Any>) {
+                    override fun onLogEvent(logEvent: LogEvent, properties: Map<LogProperty, Any>) {
 
                     }
                 }
