+++ gradle/library.versions.toml
@@ -10,7 +10,7 @@ material = "1.12.0"
 activity = "1.8.0"
 constraintlayout = "2.1.4"
 cast-framework = "21.4.0"
-bv-player = "5.0.0"
+bv-player = "5.1.0"
 
 [libraries]
 androidx-core-ktx = { group = "androidx.core", name = "core-ktx", version.ref = "coreKtx" }
